move recGroups and context types to separate modules
